requirements.txt
Django>=4.2
djangorestframework

dj-database-url
psycopg2-binary
